feat(angular): Authentication ready promise
Modified the client-side Authentication service's ready promise to
ensure we can handle the resolve externally from the service. This is
critical in the Angular route filter for changing state, in order to
check that the user has sufficient permissions to access the client-side
route.

Fixes an issue with the redirect url not getting set properly after
removing sessions.

Fixes Notification message after user signs in, to correctly show the
user's first name.

Upgraded version of passport-jwt library.